requirements.txt
astroid==2.2.5
certifi==2019.6.16
Django==2.2
django-ckeditor==5.7.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-recaptcha==2.0.4
djangorestframework==3.10.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
pipenv==2018.11.26
psycopg2==2.8.3
psycopg2-binary==2.8.3
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.2
virtualenv-clone==0.5.3
wrapt==1.11.2
